Simplified outputs in core
